+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Add mandatory pylint checks
 - Will only delete one batch of stale task at a time to improve probability of stale tasks being deleted under high load
+- Refactor to fix pylint issues
 
 ## [0.17.0] - 2023-07-04